Crate syntex_pos

The source positions and related helper functions

Note

This API is completely unstable and subject to change.

Re-exports

pub use hygiene::SyntaxContext;

pub use hygiene::ExpnInfo;

pub use hygiene::ExpnFormat;

pub use hygiene::NameAndSpan;

Modules

hygiene
Machinery for hygienic macros, inspired by the MTWT[1] paper.

symbol
An "interner" is a data structure that associates values with usize tags and allows bidirectional lookup; i.e. given a value, one can easily find the type, and vice versa.

Structs

BytePos
A byte offset. Keep this small (currently 32-bits), as AST contains a lot of them.

CharPos
A character offset. Because of multibyte utf8 characters, a byte offset is not equivalent to a character offset. The CodeMap will convert BytePos values to CharPos values as necessary.

DistinctSources

FileLines

FileMap
A single source in the CodeMap.

FileMapAndBytePos

FileMapAndLine

LineInfo

Loc
A source code location used for error reporting

LocWithOpt
A source code location used as the result of lookup_char_pos_adj

MacroBacktrace

MalformedCodemapPositions

MultiByteChar
Identifies an offset of a multi-byte character in a FileMap

MultiSpan
A collection of spans. Spans have two orthogonal attributes:

Span
Spans represent a region of code, used for error reporting. Positions in spans are absolute positions from the beginning of the codemap, not positions relative to FileMaps. Methods on the CodeMap can be used to relate spans back to the original source. You must be careful if the span crosses more than one file - you will not be able to use many of the functions on spans in codemap and you cannot assume that the length of the span = hi - lo; there may be space in the BytePos range between files.

SpanLabel

Enums

SpanLinesError

SpanSnippetError

Constants

DUMMY_SP

NO_EXPANSION

SPAN_DEBUG

Traits

Pos

Type Definitions

FileLinesResult

FileName