Struct syntex_pos::FileMap

pub struct FileMap {
    pub name: FileName,
    pub name_was_remapped: bool,
    pub crate_of_origin: u32,
    pub src: Option<Rc<String>>,
    pub start_pos: BytePos,
    pub end_pos: BytePos,
    pub lines: RefCell<Vec<BytePos>>,
    pub multibyte_chars: RefCell<Vec<MultiByteChar>>,
}

A single source in the CodeMap.

Fields
name: FileName

The name of the file that the source came from, source that doesn't originate from files has names between angle brackets by convention, e.g. <anon>

name_was_remapped: bool

True if the name field above has been modified by -Zremap-path-prefix

crate_of_origin: u32

Indicates which crate this FileMap was imported from.

src: Option<Rc<String>>

The complete source code

start_pos: BytePos

The start position of this source in the CodeMap

end_pos: BytePos

The end position of this source in the CodeMap

lines: RefCell<Vec<BytePos>>

Locations of lines beginnings in the source code

multibyte_chars: RefCell<Vec<MultiByteChar>>

Locations of multi-byte characters in the source code

Methods

impl FileMap

pub fn next_line(&self, pos: BytePos)

EFFECT: register a start-of-line offset in the table of line-beginnings. UNCHECKED INVARIANT: these offsets must be added in the right order and must be in the right places; there is shared knowledge about what ends a line between this file and parse.rs WARNING: pos param here is the offset relative to start of CodeMap, and CodeMap will append a newline when adding a filemap without a newline at the end, so the safe way to call this is with value calculated as filemap.start_pos + newline_offset_relative_to_the_start_of_filemap.

pub fn get_line(&self, line_number: usize) -> Option<&str>

get a line from the list of pre-computed line-beginnings. line-number here is 0-based.

pub fn record_multibyte_char(&self, pos: BytePos, bytes: usize)

pub fn is_real_file(&self) -> bool

pub fn is_imported(&self) -> bool

pub fn byte_length(&self) -> u32

pub fn count_lines(&self) -> usize

pub fn lookup_line(&self, pos: BytePos) -> Option<usize>

Find the line containing the given position. The return value is the index into the lines array of this FileMap, not the 1-based line number. If the filemap is empty or the position is located before the first line, None is returned.

pub fn line_bounds(&self, line_index: usize) -> (BytePos, BytePos)