Struct syntex_pos::MultiSpan

pub struct MultiSpan { /* fields omitted */ }

A collection of spans. Spans have two orthogonal attributes:

they can be primary spans. In this case they are the locus of the error, and would be rendered with ^^^.

they can have a label. In this case, the label is written next to the mark in the snippet when we render.

Methods

impl MultiSpan

fn new() -> MultiSpan

fn from_span(primary_span: Span) -> MultiSpan

fn from_spans(vec: Vec<Span>) -> MultiSpan

fn push_span_label(&mut self, span: Span, label: String)

fn primary_span(&self) -> Option<Span>

Selects the first primary span (if any)

fn primary_spans(&self) -> &[Span]

Returns all primary spans.

fn span_labels(&self) -> Vec<SpanLabel>

Returns the strings to highlight. We always ensure that there is an entry for each of the primary spans -- for each primary span P, if there is at least one label with span P, we return those labels (marked as primary). But otherwise we return SpanLabel instances with empty labels.