Struct syntex_pos::Span

pub struct Span {
    pub lo: BytePos,
    pub hi: BytePos,
    pub ctxt: SyntaxContext,
}

Spans represent a region of code, used for error reporting. Positions in spans are absolute positions from the beginning of the codemap, not positions relative to FileMaps. Methods on the CodeMap can be used to relate spans back to the original source. You must be careful if the span crosses more than one file - you will not be able to use many of the functions on spans in codemap and you cannot assume that the length of the span = hi - lo; there may be space in the BytePos range between files.

Fields
lo: BytePos hi: BytePos ctxt: SyntaxContext

Information about where the macro came from, if this piece of code was created by a macro expansion.

Methods

impl Span

pub fn end_point(self) -> Span

Returns a new span representing just the end-point of this span

pub fn next_point(self) -> Span

Returns a new span representing the next character after the end-point of this span

pub fn substitute_dummy(self, other: Span) -> Span

Returns self if self is not the dummy span, and other otherwise.

pub fn contains(self, other: Span) -> bool

pub fn source_equal(&self, other: &Span) -> bool

Return true if the spans are equal with regards to the source text.

Use this instead of == when either span could be generated code, and you only care that they point to the same bytes of source text.

pub fn trim_start(self, other: Span) -> Option<Span>

Returns Some(span), where the start is trimmed by the end of other

pub fn source_callsite(self) -> Span

Return the source span - this is either the supplied span, or the span for the macro callsite that expanded to it.

pub fn source_callee(self) -> Option<NameAndSpan>

Return the source callee.

Returns None if the supplied span has no expansion trace, else returns the NameAndSpan for the macro definition corresponding to the source callsite.

pub fn allows_unstable(&self) -> bool

Check if a span is "internal" to a macro in which #[unstable] items can be used (that is, a macro marked with #[allow_internal_unstable]).

pub fn macro_backtrace(self) -> Vec<MacroBacktrace>

pub fn to(self, end: Span) -> Span

pub fn between(self, end: Span) -> Span

pub fn until(self, end: Span) -> Span